gke-gcs-project/docker/requirements.txt
Flask==2.3.2
google-cloud-storage==2.10.0
gunicorn==20.1.0